requirements.txt
redis
Django==2.1
gunicorn==19.9.0
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
Markdown==2.6.11
Pillow==5.2.0
PyJWT==1.6.4
pytz==2018.5
six==1.11.0
requests==2.19.1